Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
langchain-huggingface==0.1.2
langchain-mistralai==0.2.10
langchain-postgres==0.0.14
langchain-text-splitters==0.3.8
langsmith==0.3.31
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistralai==1.6.0
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.16
packaging==24.2
pgvector==0.3.6
pillow==11.2.1
propcache==0.3.1
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
py-solc-x==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==1.0.5
web3>=6.11.0
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0
